+++ src/components/SearchableSingleSelect/SearchableSingleSelect.tsx
@@ -163,7 +163,6 @@ export const SearchableSingleSelect = ({
                             }
                             placeholder={i18n.t('Filter options')}
                             type="search"
-                            width={'400px'}
                         />
                     </div>
                 </div>