+++ src/lib/sectionList/listViews/sectionListViewsConfig.ts
@@ -5,6 +5,7 @@ import type { ConfigurableFilterKey } from '../filters'
 export interface ModelPropertyDescriptor {
     label: string
     path: string
+    minApiVersion?: number
 }
 
 export interface FilterDescriptor {
@@ -473,6 +474,24 @@ export const modelListViewsConfig = {
         },
         filters: {},
     },
+    relationshipType: {
+        columns: {
+            available: [
+                'bidirectional',
+                'favorite',
+                'displayFromToName',
+                'displayToFromName',
+                'fromToName',
+                'toFromName',
+                'referral',
+            ],
+        },
+        filters: {
+            default: [],
+            available: [],
+            overrideDefaultAvailable: true,
+        },
+    },
     trackedEntityType: {
         columns: {
             default: [
@@ -517,6 +536,7 @@ export const modelListViewsConfig = {
                 {
                     label: i18n.t('Trigram indexable'),
                     path: 'trigramIndexable',
+                    minApiVersion: 43,
                 },
             ],
         },