+++ src/pages/programsWip/form/ProgramFormContents.tsx
@@ -8,6 +8,8 @@ import { EnrollmentDataFormContents } from './EnrollmentDataFormContents'
 import { EnrollmentFormFormContents } from './EnrollmentFormFormContents'
 import { EnrollmentSettingsFormContents } from './EnrollmentSettingsFormContents'
 import { ProgramFormDescriptor } from './formDescriptor'
+import { ProgramCustomizationFormContents } from './ProgramCustomizationFormContents'
+import { ProgramStagesFormContents } from './ProgramStagesFormContents'
 import { SetupFormContents } from './SetupFormContents'
 
 export const ProgramFormContents = () => {
@@ -18,6 +20,9 @@ export const ProgramFormContents = () => {
             <SetupFormContents
                 name={descriptor.getSection('enrollmentDetails').name}
             />
+            <ProgramCustomizationFormContents
+                name={descriptor.getSection('programCustomization').name}
+            />
             <EnrollmentSettingsFormContents
                 name={descriptor.getSection('enrollmentSettings').name}
             />
@@ -26,6 +31,9 @@ export const ProgramFormContents = () => {
             />
             <EnrollmentFormFormContents
                 name={descriptor.getSection('enrollmentForm').name}
+            />
+            <ProgramStagesFormContents
+                name={descriptor.getSection('programStages').name}
             />
         </SectionedFormSections>
     )